ENH: Transition to use build/package.env for configuration

+++ scripts/dockcross-manylinux-build-wheels.sh
@@ -18,8 +18,14 @@
 #   scripts/dockcross-manylinux-build-module-wheels.sh cp39
 #
 script_dir=${script_dir:=$(cd $(dirname $0) || exit 1; pwd)}
-script_name=$(basename $0)
-source "${script_dir}/dockcross-manylinux-set-vars.sh"
+_ipp_dir=$(dirname $0)
+package_env_file=${_ipp_dir}/build/package.env
+if [ ! -f "${_ipp_dir}/build/package.env" ]; then
+  echo "MISSING: ${_ipp_dir}/build/package.env"
+  echo "    RUN: ${_ipp_dir}/review generate_build_environment.sh"
+  exit -1
+fi
+source "${_ipp_dir}/build/package.env"
 
 # Generate dockcross scripts
 $oci_exe run --env MANYLINUX_VERSION="${MANYLINUX_VERSION}" --rm docker.io/dockcross/manylinux${MANYLINUX_VERSION}-x64:${IMAGE_TAG} > /tmp/dockcross-manylinux-x64
